+++ src/weather.ts
@@ -1,4 +1,5 @@
 import {getBooleanSetting, getStringSetting} from './settings';
+import {deCamelCase, parseSelect} from './utils';
 
 export function dayHasChanged(game: Game, deltaInSeconds: number): boolean {
     const now = game.pf2e.worldClock.worldTime;
@@ -13,7 +14,7 @@ function getCurrentlyPlayingWeather(game: Game): StoredDocument<Playlist>[] {
         ?? [];
 }
 
-async function applyWeatherSound(game: Game, effectName: string): Promise<void> {
+async function applyWeatherSound(game: Game, effectName: WeatherEffectName): Promise<void> {
     const playlistName = `weather.${effectName}`;
 
     // stop all others playlists
@@ -22,7 +23,7 @@ async function applyWeatherSound(game: Game, effectName: string): Promise<void> 
     }
 
     // if playlist is not yet playing, start it
-    if (getCurrentlyPlayingWeather(game).length === 0) {
+    if (getCurrentlyPlayingWeather(game).length === 0 && effectName !== 'none') {
         const playlist = game.playlists?.getName(playlistName);
         if (playlist) {
             await playlist.playAll();
@@ -32,29 +33,84 @@ async function applyWeatherSound(game: Game, effectName: string): Promise<void> 
     }
 }
 
-export async function setWeather(game: Game, effectName: 'snowfall' | 'rain' | 'sunny'): Promise<void> {
-    // always persist the current one without checking for turned off weather
-    await game.settings.set('pf2e-kingmaker-tools', 'currentWeatherFx', effectName);
-    // fall back to sunny if weather is disabled
-    const eff = getBooleanSetting(game, 'enableWeather') ? effectName : 'sunny';
-    await applyWeatherSound(game, eff);
-    console.info(`Setting weather to ${eff}`);
-    if (eff === 'rain') {
-        await game.scenes?.current?.update({'weather': 'rain'});
-    } else if (eff === 'snowfall') {
-        await game.scenes?.current?.update({'weather': 'snow'});
-    } else {
-        await game.scenes?.current?.update({'weather': ''});
+const allWeatherNames = [
+    'snow',
+    'rain',
+    'sunny',
+    'leaves',
+    'rainStorm',
+    'fog',
+    'blizzard',
+    'none',
+] as const;
+type WeatherEffectName = typeof allWeatherNames[number];
+
+export async function setWeather(game: Game, effectName: WeatherEffectName): Promise<void> {
+    if (getBooleanSetting(game, 'enableWeather')) {
+        // always persist the current one without checking for turned off weather
+        await game.settings.set('pf2e-kingmaker-tools', 'currentWeatherFx', effectName);
+        // fall back to sunny if weather is disabled
+        const eff = getBooleanSetting(game, 'enableSheltered') ? effectName : 'none';
+        await applyWeatherSound(game, eff);
+        console.info(`Setting weather to ${eff}`);
+        if (eff !== 'sunny' && eff !== 'none') {
+            await game.scenes?.current?.update({'weather': eff});
+        } else {
+            await game.scenes?.current?.update({'weather': ''});
+        }
     }
 }
 
 export async function syncWeather(game: Game): Promise<void> {
-    const weather = getStringSetting(game, 'currentWeatherFx') as 'sunny' | 'snowfall' | 'rain';
+    const weather = getStringSetting(game, 'currentWeatherFx') as WeatherEffectName;
     await setWeather(game, weather);
+}
+
+export async function toggleShelterd(game: Game): Promise<void> {
+    const isEnabled = !getBooleanSetting(game, 'enableSheltered');
+    await game.settings.set('pf2e-kingmaker-tools', 'enableSheltered', isEnabled);
+    await syncWeather(game);
 }
 
 export async function toggleWeather(game: Game): Promise<void> {
     const isEnabled = !getBooleanSetting(game, 'enableWeather');
     await game.settings.set('pf2e-kingmaker-tools', 'enableWeather', isEnabled);
     await syncWeather(game);
+}
+
+export function setCurrentWeatherDialog(game: Game): void {
+    const weather = getStringSetting(game, 'currentWeatherFx') as WeatherEffectName;
+    const labels = allWeatherNames.map(name => {
+        return {label: deCamelCase(name), value: name};
+    });
+    new Dialog({
+        title: 'Set Weather',
+        content: `
+        <form class="simple-dialog-form">
+            <div>
+                <label for="km-weather">Weather</label>
+                <select name="weather" id="km-weather">
+                    ${labels.map((l) => {
+            return `<option value="${l.value}" ${weather === l.value ? 'selected' : ''}>${l.label}</option>`;
+        }).join('')}                
+                </select>
+            </div>
+        </form>
+        `,
+        buttons: {
+            roll: {
+                icon: '<i class="fa-solid fa-save"></i>',
+                label: 'Roll',
+                callback: async (html): Promise<void> => {
+                    const $html = html as HTMLElement;
+                    const weather = parseSelect($html, 'weather') as WeatherEffectName;
+                    await setWeather(game, weather);
+                },
+            },
+        },
+        default: 'roll',
+    }, {
+        jQuery: false,
+        width: 250,
+    }).render(true);
 }